tests/requirements.txt
pillow==10.0.0
aiohttp==3.8.5
python-socketio[asyncio_client]==5.9.0
